pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "megaparser"
version = "0.1.0"
description = "Multi-account Telegram parser with anti-ban rate limiting"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "kurigram==2.2.18",
    "aiogram>=3.4",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "python-dotenv>=1.0",
    "loguru>=0.7",
    "cryptography>=42.0",
    "pyyaml>=6.0",
    "aiohttp>=3.9",
    "click>=8.1",
    "python-socks[asyncio]>=2.4",
    "tgcrypto>=1.2.5",
    "opentele>=1.15",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.3",
    "mypy>=1.8",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
megaparser = "app.cli.manage:cli"
megaparser-run = "app.main:run"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]
